+++ src/ContaoManager/Plugin.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2021 Heimrich & Hannot GmbH
+ * Copyright (c) 2022 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -14,7 +14,6 @@ use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 use Contao\ManagerPlugin\Config\ConfigPluginInterface;
 use HeimrichHannot\ChoicesBundle\HeimrichHannotContaoChoicesBundle;
-use HeimrichHannot\EncoreBundle\HeimrichHannotContaoEncoreBundle;
 use HeimrichHannot\FilterBundle\HeimrichHannotContaoFilterBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 
@@ -27,10 +26,6 @@ class Plugin implements BundlePluginInterface, ConfigPluginInterface
     {
         $loadAfter = [ContaoCoreBundle::class];
 
-        if (class_exists('HeimrichHannot\EncoreBundle\HeimrichHannotContaoEncoreBundle')) {
-            $loadAfter[] = HeimrichHannotContaoEncoreBundle::class;
-        }
-
         if (class_exists('HeimrichHannot\FilterBundle\HeimrichHannotContaoFilterBundle')) {
             $loadAfter[] = HeimrichHannotContaoFilterBundle::class;
         }
@@ -42,10 +37,6 @@ class Plugin implements BundlePluginInterface, ConfigPluginInterface
 
     public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
     {
-        $loader->load('@HeimrichHannotContaoChoicesBundle/Resources/config/services.yml');
-
-        if (class_exists('HeimrichHannot\EncoreBundle\HeimrichHannotContaoEncoreBundle')) {
-            $loader->load('@HeimrichHannotContaoChoicesBundle/Resources/config/config_encore.yml');
-        }
+        $loader->load('@HeimrichHannotContaoChoicesBundle/config/services.yml');
     }
 }