+++ src/HeimrichHannotContaoChoicesBundle.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2021 Heimrich & Hannot GmbH
+ * Copyright (c) 2022 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -12,4 +12,8 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class HeimrichHannotContaoChoicesBundle extends Bundle
 {
+    public function getPath()
+    {
+        return \dirname(__DIR__);
+    }
 }